WIP - Hardware Abstraction
