+++ vent/io/config/devices.ini
@@ -1,0 +1,41 @@
+#Configuration file used to identify specific hardware devices on the Ventilator 
+
+[ADC]
+type = ADS1115
+#i2c_address = 0x48
+#i2c_bus = 1
+
+[InletValve]
+type = SolenoidValve
+pin  = 27
+form = Normally Closed
+
+[ControlValve]
+type = PWMControlValve
+pin  = 12
+form = Normally Closed
+#response_curve = path/to/file
+
+[ExpiratoryValve]
+# Change 'control' to True and change type if you plan to use the expiratory valve for more than on/off
+control = False
+type    = SolenoidValve
+pin     = 17
+form    = Normally Closed
+
+[PressureSensor]
+# Note: calibration must convert to units cm(h20)
+
+# The following config is for an Amphenol/All Sensors 20 INCH-G-P4V-MINI
+# Datasheet: https://www.mouser.com/datasheet/2/18/DS-0102_Rev_A-1499183.pdf
+type              = AnalogSensor
+channel           = 0
+offset_voltage    = 0.25
+output_span       = 4.0
+conversion_factor = 2.54*20
+
+
+[FlowSensor]
+# Note: SFM3200 is an i2c device and does not need additional calibration
+type = SFM3200
+
